requirements.txt
anthropic
gspread
google-auth
requests
